+++ app/src/main/java/com/jonas/cookly/core/data/remote/requests/auth/AuthUserRequest.kt
@@ -1,4 +1,4 @@
-package com.jonas.cookly.core.data.remote.requests
+package com.jonas.cookly.core.data.remote.requests.auth
 
 import com.google.gson.annotations.SerializedName
